+++ fireauth_templates/context_processors.py
@@ -1,8 +1,9 @@
 """
-TN-NT Templates content processor
+FI.RE Auth Templates content processor
 """
 
 # Django
+from django.conf import settings
 from django.core.handlers.wsgi import WSGIRequest
 
 
@@ -13,9 +14,12 @@ def fireauth_settings(request: WSGIRequest) -> dict:
     :return:
     """
 
-    # AA logo
+    # Template Settings
     return_value = {
-        "FIREAUTH_TEMPLATE_AA_LOGO": "/static/fireauth_templates/images/fire-coalition-32x32.png"
+        "FIREAUTH_TEMPLATE_AA_LOGO": "/static/fireauth_templates/images/fire-coalition-32x32.png",
+        "REGISTRATION_VERIFY_EMAIL": getattr(
+            settings, "REGISTRATION_VERIFY_EMAIL", True
+        ),
     }
 
     return return_value